LibWeb: Split UnsetStyleValue out of StyleValue.{h,cpp}
